move some code to tests, should not be in migrator
