+++ src/Migrator.Tests/Support/Inflector.cs
